4th attempt, I'm confident now

+++ src/TensorValues/Indexing.jl
@@ -5,8 +5,8 @@ Base.IndexStyle(::Type{<:MultiValue}) = IndexCartesian()
 
 # Necessary overloads due to wrong ::Number defaults
 lastindex(arg::MultiValue) = length(arg)
-lastindex(arg::MultiValue, d::Integer) = (@inline; size(arg, d)) # needed for 32 bit systems
-lastindex(arg::MultiValue, d::Int64) = (@inline; size(arg, d))   # method ambiguity with base
+lastindex(arg::MultiValue, d::Integer) = (@inline; size(arg, d)) # generic
+lastindex(arg::MultiValue, d::Int)     = (@inline; size(arg, d)) # method ambiguity with base
 
 # Gridap broadcast of some operation on ::MultiValue rely on Base.axes adopting
 # the Number convension (all MultiValue have axes `()` )
